+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2.0.6+2]
 * Added `BoxyDelegate.distanceToBaseline`
 * Added `BoxyDelegate.onPointerEvent`
+* Fixed constraints bug in `BoxyChild.layoutRect`
 
 ## [2.0.6+1]
 * Added `BoxyDelegate.renderSize`